# stuplus.api/swagger_output.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Stuplus API",
    "description": "Stuplus API Documentation."
  },
  "servers": [
    {
      "url": "http://212.98.224.208:25010/",
      "description": "Canli API"
    },
    {
      "url": "http://127.0.0.1:25010/",
      "description": "Local API"
    }
  ],
...
  "paths": {
...
    "/announcement/add": {
      "post": {
        "tags": [
          "Announcement"
        ],
        "description": "Add an announcement.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NullResponse"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/NullResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "coverImage": {
                    "type": "string",
                    "format": "binary"
                  },
                  "title": {
                    "type": "string"
                  },
                  "text": {
                    "type": "string"
                  },
                  "relatedSchoolIds": {
                    "type": "array"
                  },
                  "hashTags": {
                    "type": "array"
                  }
                },
                "required": [
                  "title",
                  "text",
                  "relatedSchoolIds"
                ]
              }
            }
          }
        }
      }
    },
    "/announcement/getAnnouncements": {
      "post": {
        "tags": [
          "Announcement"
        ],
        "description": "Get announcements.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnnouncementGetAnnouncementsResponse"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/AnnouncementGetAnnouncementsResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnnouncementGetAnnouncementsRequest"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/AnnouncementGetAnnouncementsRequest"
              }
            }
          }
        }
      }
    },
    "/announcement/getAnnouncement/{id}": {
      "get": {
        "tags": [
          "Announcement"
        ],
        "description": "Get announcement by id.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnnouncementGetAnnouncementResponse"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/AnnouncementGetAnnouncementResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/announcement/likeDislike": {
      "post": {
        "tags": [
          "Announcement"
        ],
        "description": "Like or dislike an announcement.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NullResponse"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/NullResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnnouncementLikeDislikeRequest"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/AnnouncementLikeDislikeRequest"
              }
            }
          }
        }
      }
    },
    "/announcement/getComments": {
      "post": {
        "tags": [
          "Announcement"
        ],
        "description": "Get comments of an announcement.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnnouncementGetCommentsResponse"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/AnnouncementGetCommentsResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnnouncementGetCommentsRequest"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/AnnouncementGetCommentsRequest"
              }
            }
          }
        }
      }
    },
    "/announcement/comment": {
      "post": {
        "tags": [
          "Announcement"
        ],
        "description": "Comment to an announcement.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NullResponse"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/NullResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnnouncementCommentRequest"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/AnnouncementCommentRequest"
              }
            }
          }
        }
      }
    },
    "/announcement/commentLikeDislike": {
      "post": {
        "tags": [
          "Announcement"
        ],
        "description": "Like or dislike a anouncement",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NullResponse"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/NullResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnnouncementCommentLikeDislikeRequest"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/AnnouncementCommentLikeDislikeRequest"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "NullResponse": {
        "type": "object",
        "properties": {
          "hasError": {
            "type": "object",
            "properties": {
              "type": {
                "type": "string",
                "example": "boolean"
              },
              "description": {
                "type": "string",
                "example": "Indicates whether the operation was successful or not."
              }
            }
          },
          "validationErrors": {
            "type": "object",
            "properties": {
              "type": {
                "type": "string",
                "example": "array"
              },
              "description": {
                "type": "string",
                "example": "List of validation errors."
              }
            }
          },
          "message": {
            "type": "object",
            "properties": {
              "type": {
                "type": "string",
                "example": "string"
              },
              "description": {
                "type": "string",
                "example": "Message describing the error."
              }
            }
          },
          "data": {
            "type": "object",
            "properties": {
              "type": {
                "type": "string",
                "example": "null"
              }
            }
          }
        },
        "xml": {
          "name": "NullResponse"
        }
      },
...
      "AnnouncementGetAnnouncementsRequest": {
        "type": "object",
        "properties": {
          "lastRecordDate": {
            "type": "string",
            "example": "2022-07-02T20:12:59.891Z"
          },
          "pageSize": {
            "type": "number",
            "example": 20
          },
          "schoolIds": {
            "type": "array",
            "example": [
              "schoolId1",
              "schoolId2"
            ],
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "pageSize"
        ],
        "xml": {
          "name": "AnnouncementGetAnnouncementsRequest"
        }
      },
      "AnnouncementGetAnnouncementsResponse": {
        "type": "object",
        "properties": {
          "hasError": {
            "type": "object",
            "properties": {
              "type": {
                "type": "string",
                "example": "boolean"
              },
              "description": {
                "type": "string",
                "example": "Indicates whether the operation was successful or not."
              }
            }
          },
          "validationErrors": {
            "type": "object",
            "properties": {
              "type": {
                "type": "string",
                "example": "array"
              },
              "description": {
                "type": "string",
                "example": "List of validation errors."
              }
            }
          },
          "message": {
            "type": "object",
            "properties": {
              "type": {
                "type": "string",
                "example": "string"
              },
              "description": {
                "type": "string",
                "example": "Message describing the error."
              }
            }
          },
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "_id": {
                  "type": "string",
                  "example": "62af8aade035eb1764400d30"
                },
                "ownerId": {
                  "type": "string",
                  "example": "62ab8a204166fd1eaebbb3fa"
                },
                "owner": {
                  "type": "object",
                  "properties": {
                    "_id": {
                      "type": "string",
                      "example": "62ab8a204166fd1eaebbb3fa"
                    },
                    "username": {
                      "type": "string",
                      "example": "ndodanli14"
                    },
                    "profilePhotoUrl": {
                      "type": "string",
                      "example": "https://stuplus-bucket.s3.amazonaws.com/public/profile_photos/1656012020219-1MFqZJCvcok.jpg"
                    },
                    "schoolId": {
                      "type": "string",
                      "example": "schoolId1"
                    },
                    "__t": {
                      "type": "string",
                      "example": "User"
                    }
                  }
                },
                "coverImageUrl": {},
                "title": {
                  "type": "string",
                  "example": "das"
                },
                "relatedSchoolIds": {
                  "type": "array",
                  "example": [
                    "string1",
                    "string2"
                  ],
                  "items": {
                    "type": "string"
                  }
                },
                "text": {
                  "type": "string",
                  "example": "dasdsa"
                },
                "isActive": {
                  "type": "boolean",
                  "example": true
                },
                "fromDate": {},
                "toDate": {},
                "__t": {
                  "type": "string",
                  "example": "Announcement"
                },
                "createdAt": {
                  "type": "string",
                  "example": "2022-06-19T20:44:29.536Z"
                },
                "updatedAt": {
                  "type": "string",
                  "example": "2022-06-19T20:44:29.536Z"
                },
                "__v": {
                  "type": "number",
                  "example": 0
                },
                "recordStatus": {
                  "type": "number",
                  "example": 1
                },
                "likeCount": {
                  "type": "number",
                  "example": 0
                },
                "commentCount": {
                  "type": "number",
                  "example": 0
                },
                "likeType": {
                  "type": "number",
                  "example": 1
                }
              }
            }
          }
        },
        "xml": {
          "name": "AnnouncementGetAnnouncementsResponse"
        }
      },
      "AnnouncementGetCommentsRequest": {
        "type": "object",
        "properties": {
          "lastRecordDate": {
            "type": "string",
            "example": "2022-07-02T20:12:59.891Z"
          },
          "pageSize": {
            "type": "number",
            "example": 20
          },
          "announcementId": {
            "type": "string",
            "example": "announcementId"
          }
        },
        "required": [
          "pageSize",
          "announcementId"
        ],
        "xml": {
          "name": "AnnouncementGetCommentsRequest"
        }
      },
...
      "AnnouncementGetAnnouncementResponse": {
        "type": "object",
        "properties": {
          "hasError": {
            "type": "object",
            "properties": {
              "type": {
                "type": "string",
                "example": "boolean"
              },
              "description": {
                "type": "string",
                "example": "Indicates whether the operation was successful or not."
              }
            }
          },
          "validationErrors": {
            "type": "object",
            "properties": {
              "type": {
                "type": "string",
                "example": "array"
              },
              "description": {
                "type": "string",
                "example": "List of validation errors."
              }
            }
          },
          "message": {
            "type": "object",
            "properties": {
              "type": {
                "type": "string",
                "example": "string"
              },
              "description": {
                "type": "string",
                "example": "Message describing the error."
              }
            }
          },
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "_id": {
                  "type": "string",
                  "example": "62af8aade035eb1764400d36"
                },
                "ownerId": {
                  "type": "string",
                  "example": "62ab8a204166fd1eaebbb3fa"
                },
                "coverImageUrl": {},
                "title": {
                  "type": "string",
                  "example": "das"
                },
                "relatedSchoolIds": {
                  "type": "array",
                  "example": [
                    "string1",
                    "string2"
                  ],
                  "items": {
                    "type": "string"
                  }
                },
                "text": {
                  "type": "string",
                  "example": "dasdsa"
                },
                "isActive": {
                  "type": "boolean",
                  "example": true
                },
                "fromDate": {},
                "toDate": {},
                "__t": {
                  "type": "string",
                  "example": "Announcement"
                },
                "createdAt": {
                  "type": "string",
                  "example": "2022-06-19T20:44:29.908Z"
                },
                "updatedAt": {
                  "type": "string",
                  "example": "2022-06-25T11:18:01.665Z"
                },
                "__v": {
                  "type": "number",
                  "example": 0
                },
                "recordStatus": {
                  "type": "number",
                  "example": 1
                },
                "score": {
                  "type": "number",
                  "example": 47
                },
                "owner": {
                  "type": "object",
                  "properties": {
                    "_id": {
                      "type": "string",
                      "example": "62ab8a204166fd1eaebbb3fa"
                    },
                    "username": {
                      "type": "string",
                      "example": "ndodanli14"
                    },
                    "profilePhotoUrl": {
                      "type": "string",
                      "example": "https://stuplus-bucket.s3.amazonaws.com/public/profile_photos/1656012020219-1MFqZJCvcok.jpg"
                    },
                    "schoolId": {
                      "type": "string",
                      "example": "schoolId1"
                    },
                    "__t": {
                      "type": "string",
                      "example": "User"
                    }
                  }
                },
                "likeCount": {
                  "type": "number",
                  "example": 22
                },
                "likeType": {
                  "type": "number",
                  "example": 2
                },
                "comments": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "_id": {
                        "type": "string",
                        "example": "62b6f0b662245f6dc443a9ae"
                      },
                      "ownerId": {
                        "type": "string",
                        "example": "62ab8a204166fd1eaebbb3fa"
                      },
                      "comment": {
                        "type": "string",
                        "example": "some comment."
                      },
                      "__t": {
                        "type": "string",
                        "example": "AnnouncementComment"
                      },
                      "owner": {
                        "type": "object",
                        "properties": {
                          "_id": {
                            "type": "string",
                            "example": "62ab8a204166fd1eaebbb3fa"
                          },
                          "username": {
                            "type": "string",
                            "example": "ndodanli14"
                          },
                          "profilePhotoUrl": {
                            "type": "string",
                            "example": "https://stuplus-bucket.s3.amazonaws.com/public/profile_photos/1656012020219-1MFqZJCvcok.jpg"
                          },
                          "__t": {
                            "type": "string",
                            "example": "User"
                          }
                        }
                      },
                      "likeCount": {
                        "type": "number",
                        "example": 22
                      },
                      "likeType": {
                        "type": "number",
                        "example": 0
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "xml": {
          "name": "AnnouncementGetAnnouncementResponse"
        }
      },
      "AnnouncementLikeDislikeRequest": {
        "type": "object",
        "properties": {
          "announcementId": {
            "type": "string",
            "example": "announcementId"
          },
          "type": {
            "type": "object",
            "properties": {
              "type": {
                "type": "string",
                "example": "enum"
              },
              "values": {
                "type": "object",
                "properties": {
                  "Dislike": {
                    "type": "number",
                    "example": 0
                  },
                  "Like": {
                    "type": "number",
                    "example": 1
                  }
                }
              }
            }
          },
          "beforeType": {
            "type": "object",
            "properties": {
              "type": {
                "type": "string",
                "example": "enum"
              },
              "description": {
                "type": "string",
                "example": "Type of the previous like/dislike."
              },
              "values": {
                "type": "object",
                "properties": {
                  "Dislike": {
                    "type": "number",
                    "example": 0
                  },
                  "Like": {
                    "type": "number",
                    "example": 1
                  },
                  "None": {
                    "type": "number",
                    "example": 2
                  }
                }
              }
            }
          }
        },
        "required": [
          "announcementId",
          "type",
          "beforeType"
        ],
        "xml": {
          "name": "AnnouncementLikeDislikeRequest"
        }
      },
      "AnnouncementCommentRequest": {
        "type": "object",
        "properties": {
          "announcementId": {
            "type": "string",
            "example": "announcementId"
          },
          "comment": {
            "type": "string",
            "example": "some comment."
          }
        },
        "required": [
          "announcementId",
          "comment"
        ],
        "xml": {
          "name": "AnnouncementCommentRequest"
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}